WebServer v2

+++ Arduino/Sonar.cpp
@@ -7,7 +7,7 @@
 
 #include "Sonar.h"
 
-Sonar::Sonar(uint8_t pinSonarRead, uint8_t pinSonarWrite, int maxDistance) : pinSonarRead(pinSonarRead), pinSonarWrite(pinSonarWrite), maxDistance(maxDistance) {
+Sonar::Sonar(uint8_t pinSonarRead, uint8_t pinSonarWrite, uint16_t maxDistance) : pinSonarRead(pinSonarRead), pinSonarWrite(pinSonarWrite), maxDistance(maxDistance) {
 //    pinMode(pinSonar, INPUT);
 //    ultrasonic(pinSonar, 11, maxDistance);
     ultrasonic = new Ultrasonic(pinSonarRead, pinSonarWrite);
@@ -16,6 +16,6 @@ Sonar::Sonar(uint8_t pinSonarRead, uint8_t pinSonarWrite, int maxDistance) : pin
 Sonar::~Sonar() {
 }
 
-int Sonar::readDistance(){
+uint16_t Sonar::readDistance(){
     return ultrasonic->convert(ultrasonic->timing(), Ultrasonic::CM);
 }
